requirements.txt
# ============================================
# DEPENDENCIAS - Sistema de Gestión de Documentos Seguros
# ============================================

# Framework base
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0

# Seguridad y autenticación
pyotp==2.9.0
qrcode==7.4.2
pillow

# Utilidades
python-dotenv==1.0.0
Werkzeug==2.3.7

# Base de datos (opcional para producción)
# psycopg2-binary==2.9.7  # PostgreSQL

# Desarrollo y testing
pytest==7.4.3
pytest-flask==1.3.0

# Sistema de caché avanzado
Flask-Caching==2.1.0
redis==5.0.1

# Rate limiting granular  
Flask-Limiter==3.5.0

# Compresión HTTP automática
Flask-Compress==1.14

# Migración y optimización de DB
Flask-Migrate==4.0.5

# Monitoreo y métricas
#Flask-APM==0.1.1
psutil==5.9.6

# Logging avanzado
colorlog==6.8.0

# Utilidades existentes
python-dotenv==1.0.0
Werkzeug==2.3.7
